change dtype to float

+++ basic_nn_MNIST/main.py
@@ -76,9 +76,9 @@ def loss_fn(model, features, config, training):
     mean_classification_loss = tf.reduce_mean(classification_loss) 
 
     accuracy = tf.reduce_mean(
-        tf.where(tf.equal(label, tf.argmax(model_output, axis=-1)),
-                    tf.ones_like(label, dtype=tf.int64),
-                    tf.zeros_like(label, dtype=tf.int64)))
+        tf.where(tf.equal(tf.cast(label, dtype=tf.float32), tf.argmax(model_output, axis=-1)),
+                    tf.ones_like(label, dtype=tf.float32),
+                    tf.zeros_like(label, dtype=tf.float32)))
 
     return mean_classification_loss, l2_regularizer, accuracy, classification_loss
